+++ dolang-runtime/src/object/float.rs
@@ -379,7 +379,7 @@ fn coerce_to_f64<'v, 's>(value: &Value<'v>, strand: &mut Strand<'v, 's>) -> Resu
         match value.to_prim(strand)? {
             Prim::Int(v) => Ok(v as f64),
             Prim::F64(v) => Ok(v),
-            Prim::Bool(v) => Ok(v as i64 as f64),
+            Prim::Bool(v) => Ok(v as i32 as f64),
             Prim::Nil => Err(Error::type_error(strand, "float: `nil` can't be converted")),
         }
     }
